RAG/requirements.txt
python-dotenv
numpy
PyMuPDF
langchain-openai
langchain
langchain-core
langchain-community
openai